# Allocate the shards only once while reading WAL