The first level of the 2D slice need not be allocated multiple times as only the 2nd level slice is moved around.
Benchmark (the `ns/op` and `B/op` seem to jump quite a bit with repeated tests):
```
benchmark                                                                    old ns/op      new ns/op      delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=100000-8     6590119060     6156305331     -6.58%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=100-8      1245573574     1315054689     +5.58%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=10000-8     6200068172     5907548411     -4.72%

benchmark                                                                    old allocs     new allocs     delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=100000-8     7079462        6079513        -14.12%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=100-8      2466872        2464654        -0.09%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=10000-8     6429524        6329664        -1.55%

benchmark                                                                    old bytes     new bytes     delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=100000-8     830975392     638984648     -23.10%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=100-8      360086352     350880864     -2.56%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=10000-8     798231440     779473256     -2.35%
```

As expected the savings are when there are more records with samples.